ARM: proc-v7.S: remove HARVARD_CACHE preprocessor guards
On v7, we use the same cache maintenance instructions for data lines
as for unified lines. This was not the case for v6, where HARVARD_CACHE
was defined to indicate the L1 cache topology.

This patch removes the erroneous compile-time check for HARVARD_CACHE in
proc-v7.S, ensuring that we perform I-side invalidation at boot.

Reported-by: Shawn Guo <shawn.guo@xxxxxxxxxx>
Acked-by: Catalin Marinas <Catalin.Marinas@xxxxxxx>
Signed-off-by: Will Deacon <will.deacon@xxxxxxx>